Add location and GPS permission checks for pams survey and construction monitoring level done

+++ lib/presentation/construction_monitoring/cm_sub_prj_page.dart
@@ -10,6 +10,7 @@ import 'package:lf_survey/constants/app_dimens.dart';
 import 'package:lf_survey/constants/app_images.dart';
 import 'package:lf_survey/constants/app_text_style.dart';
 import 'package:lf_survey/constants/snackbar_helper.dart';
+import 'package:lf_survey/constants/utils.dart';
 import 'package:lf_survey/cubit/construction_monitering/cm_sub_prj/cm_sub_prj_cubit.dart';
 import 'package:lf_survey/cubit/construction_monitering/cm_sub_prj/cm_sub_prj_state.dart';
 import 'package:lf_survey/model/construction_monitoring/cm_building_response.dart';
@@ -85,7 +86,8 @@ class _CMSubPrjPageState extends State<CMSubPrjPage> {
     return Scaffold(
       backgroundColor: AppColors.appBg,
       appBar: CustomAppBar(
-        title: "Sub Projects",
+        // title: "Sub Projects",
+        title: "Wings",
         actions: [
           Padding(
             padding: const EdgeInsets.only(right: 12.0),
@@ -221,6 +223,14 @@ class _CMSubPrjPageState extends State<CMSubPrjPage> {
                                 padding: const EdgeInsets.only(bottom: 8.0),
                                 child: InkWell(
                                   onTap: () async {
+                                    final locPermission = await Utils.checkLocationAndGpsPermission(context);
+                                    if (!context.mounted) return;
+                                    if (!locPermission) {
+                                      CustomSnackHelper.errorToast(
+                                        message: "Please enable location permission and GPS",
+                                      );
+                                      return;
+                                    }
                                     wingData.projectId = widget.prjDatum.projectId;
                                     await context.pushNamed(AppRoutesName.cmSurveyPage, extra: {"wingData": wingData});
                                     cmSubPrjCubit.getSurvey(projectId: widget.prjDatum.projectId!);
@@ -268,7 +278,17 @@ class _CMSubPrjPageState extends State<CMSubPrjPage> {
                                               // if (wingData.createdWingId != null && wingData.submitStatus != true)
                                               if (wingData.submitStatus != true)
                                                 GestureDetector(
-                                                  onTap: () {
+                                                  onTap: () async {
+                                                    final locPermission = await Utils.checkLocationAndGpsPermission(
+                                                      context,
+                                                    );
+                                                    if (!context.mounted) return;
+                                                    if (!locPermission) {
+                                                      CustomSnackHelper.errorToast(
+                                                        message: "Please enable location permission and GPS",
+                                                      );
+                                                      return;
+                                                    }
                                                     // Delete wing which is created by user
                                                     if (wingData.createdWingId != null) {
                                                       CutsomAlertDialogues.deleteDialogue(
@@ -366,6 +386,16 @@ class _CMSubPrjPageState extends State<CMSubPrjPage> {
                                               onTap: wingData.submitStatus == true
                                                   ? null
                                                   : () async {
+                                                      final locPermission = await Utils.checkLocationAndGpsPermission(
+                                                        context,
+                                                      );
+                                                      if (!context.mounted) return;
+                                                      if (!locPermission) {
+                                                        CustomSnackHelper.errorToast(
+                                                          message: "Please enable location permission and GPS",
+                                                        );
+                                                        return;
+                                                      }
                                                       wingSurvey.isEmpty || wingImage.isEmpty
                                                           ? CutsomAlertDialogues.dataAlertDialogue(context: context)
                                                           : wingUnsyncSurvey.isNotEmpty || wingUnsyncImage.isNotEmpty
